Make RelevancePicker a FormField everywhere.
